Add files via upload
The code above downloads two PDF files from remote URLs using Python's requests library, and saves them to a local directory called output.

The urls variable is a list that contains the URLs of the two PDF files to be downloaded. The output_dir variable specifies the directory where the downloaded files will be saved.

A loop is then created to iterate over each URL in the urls list. For each URL, an HTTP GET request is made using the requests.get() method, and the response is stored in the r variable.

If the HTTP response status code is 200 (which indicates a successful request), the code uses the os.path.basename() method to extract the file name from the URL, and the os.path.join() method to create a full file path by concatenating the output_dir and file name.

The code then opens the file in binary write mode ('wb') using the open() function, and writes the contents of the HTTP response to the file using the write() method.

There is a small error in the line where the file_path variable is defined. It should use the url variable instead of urls, like this:
file_path = os.path.join(output_dir, os.path.basename(url))

This will ensure that the correct file name is used for each URL in the loop.

Overall, this code downloads two PDF files from remote URLs and saves them to a local directory, which can be useful for automated data scraping or other similar tasks.